feat: moved functions to utils
