+++ src/core/errors/index.ts
@@ -12,8 +12,8 @@ export { NotFoundError } from './not-found-error';
 export { StateError } from './state-error';
 export { TransactionError } from './transaction-error';
 export { TransactionPrecheckError } from './transaction-precheck-error';
+export { ValidationError } from './validation-error';
 export {
   formatZodIssueLine,
   formatZodIssuesForMessage,
-  ValidationError,
-} from './validation-error';
+} from '@/core/utils/format-zod-issues';
