+++ src/core/errors/validation-error.ts
@@ -1,17 +1,11 @@
 import type { ZodError } from 'zod';
 
-import { type $ZodIssue, toDotPath } from 'zod/v4/core';
+import {
+  formatZodIssueLine,
+  formatZodIssuesForMessage,
+} from '@/core/utils/format-zod-issues';
 
 import { CliError } from './cli-error';
-
-export function formatZodIssueLine(issue: $ZodIssue): string {
-  const path = issue.path.length === 0 ? '(root)' : toDotPath(issue.path);
-  return `${path}: ${issue.message}`;
-}
-
-export function formatZodIssuesForMessage(zodError: ZodError): string {
-  return zodError.issues.map((i) => `  - ${formatZodIssueLine(i)}`).join('\n');
-}
 
 export class ValidationError extends CliError {
   static readonly CODE = 'VALIDATION_ERROR';
